+++ javascript/examples/processTimeout.ts
@@ -1,37 +1,36 @@
 /*
- * Demonstrates how to stop an iterative job using the abort signal.
+ * Demonstrates perform timeout using the `performTimeout` option on JobDef.
  *
  * To Test:
  *
- * (1) Run script: npx ts-node examples/abortSignal.ts
+ * (1) Run script: npx ts-node examples/processTimeout.ts
  * (2) Publish a message: nats pub ORDERS someText
- * (3) Crl-C the script in the middle of the 1-5 log messages.
+ * (3) Wait one second to see the message be rejected with a timeout error.
  *
  * Requires NATS to be running.
  */
 import { JsMsg } from 'nats'
 import { setTimeout } from 'node:timers/promises'
 import { jobProcessor } from '../src/jobProcessor'
-import { Context } from '../src/types'
 
 const def = {
   stream: 'ORDERS',
-  async perform(msg: JsMsg, { signal }: Context) {
+  async perform(msg: JsMsg) {
     console.log(`Started ${msg.info.streamSequence}`)
-    for (let i = 0; i < 5; i++) {
-      await setTimeout(1000)
-      console.log(`Iteration ${i + 1} of 5`)
-      if (signal.aborted) {
-        return
-      }
-    }
+    await setTimeout(5000)
     console.log(`Completed ${msg.info.streamSequence}`)
   },
+  performTimeout: 1000,
+  numAttempts: 1,
 }
 const run = async () => {
   const processor = await jobProcessor()
   // Start processing messages
   const ordersJob = processor.start(def)
+  processor.emitter.on('start', console.info)
+  processor.emitter.on('complete', console.info)
+  processor.emitter.on('error', console.error)
+
   // Gracefully handle signals
   const shutDown = async () => {
     await ordersJob.stop()